refactor(web): hoist shared modules out of domains to resolve 4 import cycles (LUM-1630)
Cross-domain audit found 8 bidirectional cycles. Three clean hoists kill
four of them (and every longer cycle that depended on those edges):

- turn-selectors operates purely on messaging/turn-store.
  Move chat/utils/turn-selectors → messaging/turn-selectors.
  Kills chat ↔ messaging (the back-edge was a test importing
  turn-selectors).

- local-settings is a generic localStorage helper, consumed by settings,
  onboarding, and nudges. Move settings/local-settings → lib/local-settings.
  Kills onboarding ↔ settings and nudges ↔ settings.

- file-markdown is a markdown-rendering primitive used by intelligence,
  workspace, and chat. Move intelligence/components/file-markdown →
  components/file-markdown.
  Kills chat ↔ intelligence (the chat-side back-edge was the only
  chat→intelligence import; the unidirectional intel→chat dependency is
  allowed by the convention).

Allowlist updated for the new top-level homes: removed entries for files
that no longer cross domains, added chat → messaging for chat-page.tsx
(which now reaches turn-selectors via messaging).

Remaining cycles, deferred to follow-ups since they need deeper
restructuring than a hoist:
- chat ↔ conversations (chat utils that conversations depends on;
  41 edges combined)
- chat ↔ subagents (subagents imports chat/api/event-types — a 738-line
  shared types file)
- chat ↔ interactions (interactions/interaction-store imports
  chat/types/chat-ui-types, which itself depends on event-types)
- chat ↔ settings (chat handler imports a settings routing helper;
  settings imports four chat APIs)

ESLint enforcement (LUM-1632) stays deferred until those land.
