+++ apps/web/src/domains/settings/pages/integrations-page.tsx
@@ -31,7 +31,7 @@ import { routes } from "@/utils/routes.js";
 import {
   getLocalSetting,
   setLocalSetting,
-} from "@/domains/settings/local-settings.js";
+} from "@/lib/local-settings.js";
 
 const BANNER_STORAGE_KEY = "integrations.bannerDismissed";
