+++ apps/web/src/domains/settings/pages/voice-page.tsx
@@ -16,7 +16,7 @@ import { SettingsCard } from "@/domains/settings/components/settings-card.js";
 import {
   getLocalSetting,
   setLocalSetting,
-} from "@/domains/settings/local-settings.js";
+} from "@/lib/local-settings.js";
 import {
   LS_PTT_ACTIVATION_KEY,
   activatorDisplayName,
