+++ apps/web/src/domains/workspace/components/workspace-file-viewer.tsx
@@ -21,7 +21,7 @@ import { useCallback, useEffect, useRef, useState, type ReactNode } from "react"
 
 import { Button } from "@vellum/design-library/components/button";
 import { client } from "@/generated/api/client.gen.js";
-import { FileMarkdown, isMarkdown } from "@/domains/intelligence/components/file-markdown.js";
+import { FileMarkdown, isMarkdown } from "@/components/file-markdown.js";
 import { isJson, prettifyJson } from "@/domains/workspace/utils/file-json.js";
 import { formatFileSize } from "@/domains/workspace/utils/format-file-size.js";
